pex_binary: Pass interpreter attribute as --python-shebang
The pex_binary interpreter attribute is documented as "Path to the
python interpreter the pex should to use in its shebang line."
However, it was not being passed when building the PEX, so it did not
take effect.

It is getting passed as PEX_PYTHON, which changes the interpreter used
for *that invocation* of a given pex. This may have been a recent
change: https://github.com/pantsbuild/pex/issues/53

interpreter_test.py: Add a test that reads the files and verifies
    that the interpreter is set correctly.